added code to check if the code is already used in treatment.php

+++ minichat_treatment.php
@@ -9,11 +9,29 @@ $message = htmlspecialchars($_POST['message']);
 $code = (int) $_POST['code'];
 $supcode = (int) $_POST['supcode'];
 
+// Variables to check if the code is already used
+$testcode = $bdd->query('SELECT pseudo FROM minichat WHERE code = ' . $code);
+$donnees = $testcode->fetch();
+
 // Prepare and insert the data
-$request = $bdd->prepare('INSERT INTO minichat(pseudo, message, code) VALUES(?, ?, ?)');
-$request->execute(array($pseudo, $message, $code));
+if (!empty($pseudo) && !empty($message) && !empty($code)) {
+
+  if ($donnees == false) {
+    $request = $bdd->prepare('INSERT INTO minichat(pseudo, message, code) VALUES(?, ?, ?)');
+    $request->execute(array($pseudo, $message, $code));
+  }
+  else {
+    header('Location: minichat.php?code=1');
+    exit();
+  }
+}
+
+// If a suppression code is sent check if it matched the database and delete the message
+if (!empty($supcode)) {
 
 $bdd->exec('DELETE FROM minichat WHERE code = ' . $supcode);
+
+}
 
 header('Location: minichat.php');
